requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==4.10.1
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn
python-multipart

-e .